+++ android-pubsub/src/main/java/com/pj/pubsub/MessageHolder.kt
@@ -1,0 +1,9 @@
+package com.pj.pubsub
+
+import com.pj.pubsub.extensions.Message
+
+interface MessageHolder {
+
+    val message : Message
+    fun giveBack(message: Message)
+}