+++ android-pubsub/src/main/java/com/pj/pubsub/MessageMediator.kt
@@ -1,0 +1,9 @@
+package com.pj.pubsub
+
+import com.pj.pubsub.extensions.Message
+
+interface MessageMediator {
+    fun register(node: ReceivablePublisher)
+    fun publish(message: Message, tag: Tag, publisher: Publisher)
+    fun giveBack(message: Message, giveBacked: Receivable)
+}