+++ android-pubsub/src/main/java/com/pj/pubsub/unity/NativeBridgeCallback.java
@@ -1,0 +1,5 @@
+package com.pj.pubsub.unity;
+
+public interface NativeBridgeCallback{
+    void onReceive(byte[] data);
+}