+++ androidBridgeCore/src/main/java/com/pj/core/unity/NativeBridgeCallback.java
@@ -1,5 +1,0 @@
-package com.pj.core.unity;
-
-public interface NativeBridgeCallback{
-    void onReceive(byte[] data);
-}