+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 # Changelog
 
 ## [Unreleased]
-
+### Changed
+- Rename package com.pj.core to com.pj.pubsub
+- Rename module name androidBridgeCore to android-pubsub
+- Rename MessageHandler's method
+- Rename MessageHandler to Messenger.
+### Removed
+- Delete MessageCollector.
 
 ## [0.0.1] - 2024-02-29